Fjernet static i tester ryddet opp kode slik at den ser mer ryddig ut

+++ app/src/main/java/no/hvl/dat153/oblig2/main/activities/DatabaseActivity.java
@@ -13,6 +13,7 @@ import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.Toast;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -24,16 +25,11 @@ import no.hvl.dat153.oblig2.main.ui.StudentViewModel;
 
 public class DatabaseActivity extends AppCompatActivity {
 
-
-
-    private RecyclerView.LayoutManager mLayoutManager;
     private Button buttonDelete;
     private EditText editTextDelete;
     private StudentViewModel mViewModel;
     private StudentListAdapter adapter;
-
-    //static her ikke bra
-    public static ArrayList<Student> pList;
+    public ArrayList<Student> pList;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -43,23 +39,23 @@ public class DatabaseActivity extends AppCompatActivity {
         mViewModel = new ViewModelProvider(this).get(StudentViewModel.class);
         buttonDelete = findViewById(R.id.button_delete);
         editTextDelete = findViewById(R.id.edittext_delete);
-
         observerSetup();
         recyclerSetup();
         deleteItemOnclick();
     }
 
-
     public void deleteItemOnclick() {
         buttonDelete.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                mViewModel.deleteStudentId(Integer.parseInt(editTextDelete.getText().toString()));
+                if (editTextDelete.getText().toString().equals("")) {
+                    Toast.makeText(DatabaseActivity.this, "Skriv inn ID før du Sletter!", Toast.LENGTH_SHORT).show();
+                } else {
+                    mViewModel.deleteStudentId(Integer.parseInt(editTextDelete.getText().toString()));
+                }
             }
         });
     }
-
-
 
     public void goAdd(View View) {
         Intent i = new Intent(this, AddActivity.class);
@@ -88,20 +84,28 @@ public class DatabaseActivity extends AppCompatActivity {
     }
 
     public void deleteItem(int id) {
-    buttonDelete.setOnClickListener(new View.OnClickListener() {
-        @Override
-        public void onClick(View view) {
-            mViewModel.deleteStudentId(id);
-        }
-    });
-}
+        buttonDelete.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                mViewModel.deleteStudentId(id);
+            }
+        });
+    }
 
-        private void recyclerSetup() {
+    private void recyclerSetup() {
         RecyclerView recyclerView;
         adapter = new StudentListAdapter(R.layout.student_list_item);
         recyclerView = findViewById(R.id.recyclerView);
-
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
         recyclerView.setAdapter(adapter);
     }
+
+    public int getFirstPersonId() {
+        return pList.get(0).getId();
+    }
+
+    public int getDatabaseSize() {
+        return pList.size();
+    }
+
 }
